rename extension settings provider
