+++ src/DetailsView/components/details-view-overlay/settings-panel/settings/extension-settings-provider.ts
